--- src/internal/mod.rs
@@ -1,13 +1,0 @@
-#[macro_use]
-mod macros;
-
-pub mod builder;
-pub mod cabinet;
-pub mod checksum;
-pub mod consts;
-pub mod ctype;
-pub mod datetime;
-pub mod file;
-pub mod folder;
-pub mod mszip;
-mod string;
